# contracts/openapi/requests.yaml
openapi: 3.1.0
info:
  title: Requests API
  version: 1.1.0
  description: External REST contract (OpenAPI 3.1) for the request lifecycle golden path.
servers:
  - url: http://localhost:8081
    description: local gateway

paths:
  /requests:
    get:
      operationId: listRequests
      summary: List requests
      description: Lists requests for the operator workflow, optionally filtered by status or subject.
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/RequestStatus'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        '200':
          description: Requests visible to the caller
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestList'
        '400':
          description: Invalid filter
        '401':
          description: Unauthenticated
    post:
      operationId: createRequest
      summary: Create a request
      description: Creates a request for the authenticated subject and emits RequestCreated.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequest'
      responses:
        '201':
          description: Request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
        '400':
          description: Invalid input
        '401':
          description: Unauthenticated

  /requests/{id}:
    get:
      operationId: getRequest
      summary: Get a request
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
        '401':
          description: Unauthenticated
        '404':
          description: Request not found

  /requests/{id}/status:
    patch:
      operationId: updateRequestStatus
      summary: Update request status
      description: Applies an allowed lifecycle transition and emits RequestStatusChanged.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRequestStatus'
      responses:
        '200':
          description: Updated request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
        '400':
          description: Invalid input
        '401':
          description: Unauthenticated
        '403':
          description: Operator role required
        '404':
          description: Request not found
        '409':
          description: Illegal lifecycle transition

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RequestStatus:
      type: string
      enum: [created, triaged, in_progress, resolved, closed, cancelled]
    CreateRequest:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 4000
    UpdateRequestStatus:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/RequestStatus'
    Request:
      type: object
      additionalProperties: false
      required: [id, title, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
        status:
          $ref: '#/components/schemas/RequestStatus'
        created_at:
          type: string
          format: date-time
    RequestList:
      type: object
      additionalProperties: false
      required: [requests]
      properties:
        requests:
          type: array
          items:
            $ref: '#/components/schemas/Request'